login and fetch user

+++ src/app/components/Trace/trace-affich/trace-affich.component.ts
@@ -1,4 +1,3 @@
-
 import { Component, OnInit } from '@angular/core';
 import { TraceService } from '../../../services/trace.service';
 import { Trace } from '../../../models/trace';
@@ -6,20 +5,19 @@ import { Trace } from '../../../models/trace';
 @Component({
   selector: 'app-trace-affich',
   templateUrl: './trace-affich.component.html',
-  styleUrl: './trace-affich.component.css'
+  styleUrls: ['./trace-affich.component.css']
 })
-
-
-
 export class TraceAffichComponent implements OnInit {
   traces: Trace[] = [];
   filteredTraces: Trace[] = [];
   searchTerm: string = '';
+  userRole: string | null = null;
 
   constructor(private traceService: TraceService) {}
 
   ngOnInit(): void {
     this.loadTraces();
+    this.userRole = localStorage.getItem('userRole');
   }
 
   loadTraces(): void {
@@ -44,13 +42,30 @@ export class TraceAffichComponent implements OnInit {
   }
 
   deleteTrace(id: number): void {
-    this.traceService.deleteTrace(id).subscribe(
-      () => {
-        this.loadTraces();
-      },
-      (error) => {
-        console.error('Error deleting trace', error);
-      }
-    );
+    if (this.userRole === 'Admin') {
+      this.traceService.deleteTrace(id).subscribe(
+        () => {
+          this.loadTraces();
+        },
+        (error) => {
+          console.error('Error deleting trace', error);
+        }
+      );
+    }
+  }
+
+  // Function to determine if the 'Create trace' button should be enabled
+  canCreateTrace(): boolean {
+    return this.userRole === 'Admin' || this.userRole === 'Visiteur';
+  }
+
+  // Function to determine if the 'Edit' button should be enabled
+  canEditTrace(): boolean {
+    return this.userRole === 'Admin' || this.userRole === 'Visiteur';
+  }
+
+  // Function to determine if the 'Delete' button should be enabled
+  canDeleteTrace(): boolean {
+    return this.userRole === 'Admin';
   }
 }
